Extract additional functions to utils
