+++ src/Components/Issues.jsx
@@ -4,44 +4,19 @@ import IssuesWrapper from './IssuesWrapper';
 import IssueDetails from './IssueDetails';
 import FilteringCriteria from './FilteringCriteria';
 import IssuesList from './IssuesList';
-
-const retrieveLabelsDescriptions = (issues) => {
-  return issues.map((issue) => {
-    const labelsDescriptions = issue.labels.map(({ description }) => description);
-    return { ...issue, labels: labelsDescriptions };
-  });
-};
-
-const filterIssuesByCriteria = (criteria, issues) => {
-  return issues.filter((issue) => {
-    return issue.labels.some((label) => label.includes(criteria.labelFilter));
-  })
-   .filter((issue) => {
-     if (!criteria.assigneeFilter) {
-       return true;
-     }
-     return issue.assignee?.startsWith(criteria.assigneeFilter);
-   });
-};
-
-const sortIssues = (order, issues) => {
-  const compareFunctions = {
-    'none': () => 0,
-    'desc': (a, b) => new Date(a.updated_at) - new Date(b.updated_at),
-    'asc': (a, b) => new Date(b.updated_at) - new Date(a.updated_at)
-  };
-  return issues.sort(compareFunctions[order]);
-};
+import { retrieveLabelsDescriptions, filterIssuesByCriteria, sortIssues } from '../utils';
 
 const Issues = ({ issues }) => {
   const initialRenderInfo = { render: "list" };
   const [renderInfo, setRenderInfo] = useState(initialRenderInfo);
+
   const initCriteria = {
     labelFilter: '',
     assigneeFilter: '',
     sortOrder: 'none',
   };
   const [currentCriteria, setCriteria] = useState(initCriteria);
+  
   const showIssueDetails = (issue) => setRenderInfo({ render: "details", item: issue });
   const showIssuesList = () => setRenderInfo(initialRenderInfo);
